Form submits to js, not server
Default behavior of submit overridden by adding event listener to
the submit button `event.preventDefault()`.

Form data now goes to javascript by manipulating the dom to grab
data currently typed into input boxes via getElementById("id").value.

One problem is that the default value of a checkbox is to be "on", and
the user clicking a box does not actually count as changing the value,
so a conditional statement is written to see if the checkbox is
currently checked or unchecked by getElementById("read-input").checked