Add delete button to each tr
These delete buttons added to the end of each table row successfully
delete the object in the array they are associated with and will
automatically update the main table.

This was accomplished by associating a each delete button with its
corresponding array object with a variable called "domID" which
will increment when the buildTable function is called. Each remove
button will remove only the object it was tied to in that row, as it
deletes them by id.

Example: deleteBook("btn-3") will remove the book object at array
index 3 in myLibrary.