Click to toggle read/unread
User can now click on read or unread in the table to change the
status of read in the object and update the table instantly.

This is done by using if-statements in the buildTable() fucntion
that test for booleans, as 'read' is the only key that can have a
boolean for a value. The if-blocks use continue to override the
default td.appendChild behavior and adds an event listener for
toggleRead(). However, only toggleRead() actually changes the
object within the array.