Validate user input and provide feedback
validateUserFeedback() now acts as a custom validator for user input
before sending data to the getFormData() function. This ensures that
a user must put in the correct type of data before an object will be
created and the table updated.

User feedback is provided by hidden divs having their display turned
on from conditional if-statements depending on if they put in the
correct data or not.

If and only if a fully validated form is filled out, will getFormData()
actually execute.