Table no longer disappears forever
Debugged an issue where the table was disappearing forever once all
books were deleted. This is because the buildTable function was
trying to build a table from an empty object and threw an error.

Solved issue by modifying the deleteBook() function to check the
length of the myLibrary array. If the length is 1 or less, it will
not allow the user to delete the object. This effectively makes it
so that the user cannot delete their last remaining book.